Rename namespace: Cloudflare --> Api
