+++ src/Sunny.php
@@ -20,7 +20,7 @@ namespace TypistTech\Sunny;
 
 use TypistTech\Sunny\Ads\I18nPromoter;
 use TypistTech\Sunny\Ads\ReviewNotice;
-use TypistTech\Sunny\Cloudflare\Admin as CloudflareAdmin;
+use TypistTech\Sunny\Api\Admin as ApiAdmin;
 use TypistTech\Sunny\REST\Controllers\Posts\Caches\DeleteController as PostsCachesDeleteController;
 use TypistTech\Sunny\REST\Controllers\Posts\RelatedUrls\IndexController as PostsRelatedUrlsIndexController;
 use TypistTech\Sunny\Vendor\League\Container\Container;
@@ -67,7 +67,7 @@ final class Sunny implements LoadableInterface
         $loadables = [
             __CLASS__,
             Admin::class,
-            CloudflareAdmin::class,
+            ApiAdmin::class,
             I18n::class,
             I18nPromoter::class,
             PostsCachesDeleteController::class,
